requirements.txt
# Dependências principais
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0

# Utilitários
datetime
json
logging

# Para desenvolvimento
pylint>=2.11.0